TASK 2/requirements.txt
Flask
pytesseract
pillow
opencv-python
